Refactor using new interface style
The MEM ecosystem now has a new interface style, and the client application has been updated to match.
It is important to note that the compositional relationships follow the control flow of a message request, so the client owns the request interface (which is the reverse of the server, where the server is owned by the replying interface).